chore: make xcode project references back to group type
